Added Guide 6
Attention: I had no concern about optimizing the perfomance of any function. For example, the function SeqListIsPalindromic (and other functions), for maximum perfomance, should iterate directly on the list (ie, instead of calling the function SeqListGetElement). However, this would affect the modularization of this implementation of a Doubly Linked List, and thus I opted for a less efficient solution.

Note 1: a recursive version of many of the functions of both data structures would seem more obvious but they're easily convertable to the iterative equivalent...

Note 2: Guide 7 is ready but I will not upload it for now since not everyone has submited it.